Create README.md
Read a text file inputted via argument to the code, perform the below given sequence of actions and write it to new file. 
a.	Print the number of words having prefix with “To” in the input file.
b.	Print the number of words ending with “ing” in the input file.
c.	Print the word that was repeated maximum number of times.
d.	Print the palindrome present in the file.
e.	Convert all words into unique list and print in command line
f.	Create a Word dict with Key as counter index and value as the words present in file and print them on screen
     g.	Write new file with following actions
              i.	Split the words based on the vowels
              ii.	Capitalize 3rd letter of every word 
             iii.	Capitalize all characters of every 5th word in the file.
             iv.	Use – in place of blank space
             v.	Use ; (semi-colon) for new line.
             vi.	Output file name should be generated with unique name.